refactor: 싱글톤 객체로 전환
