+++ src/main/kotlin/step2/domain/operation/command/implement/DivisionOperationCommand.kt
@@ -1,7 +1,8 @@
-package step2.domain.operation.command
+package step2.domain.operation.command.implement
 
+import step2.domain.operation.command.OperationCommand
 import step2.domain.operation.model.InputOperationCommand
 
-class DivisionOperationCommand : OperationCommand {
+object DivisionOperationCommand : OperationCommand {
     override fun operate(input: InputOperationCommand) = input.firstNumber / input.secondNumber
 }
