+++ src/main/kotlin/step2/domain/operation/command/implement/MultiplicationOperationCommand.kt
@@ -1,0 +1,8 @@
+package step2.domain.operation.command.implement
+
+import step2.domain.operation.command.OperationCommand
+import step2.domain.operation.model.InputOperationCommand
+
+object MultiplicationOperationCommand : OperationCommand {
+    override fun operate(input: InputOperationCommand) = input.firstNumber * input.secondNumber
+}
