+++ src/test/kotlin/step2/domain/operator/command/OperationCommandTest.kt
@@ -4,12 +4,12 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
-import step2.domain.operation.command.AdditionOperationCommand
-import step2.domain.operation.command.DefaultOperationCommand
-import step2.domain.operation.command.DivisionOperationCommand
-import step2.domain.operation.command.MultiplicationOperationCommand
 import step2.domain.operation.command.OperationCommand
-import step2.domain.operation.command.SubtractionOperationCommand
+import step2.domain.operation.command.implement.AdditionOperationCommand
+import step2.domain.operation.command.implement.DefaultOperationCommand
+import step2.domain.operation.command.implement.DivisionOperationCommand
+import step2.domain.operation.command.implement.MultiplicationOperationCommand
+import step2.domain.operation.command.implement.SubtractionOperationCommand
 import step2.domain.operation.model.InputOperationCommand
 
 class OperationCommandTest {
@@ -22,7 +22,7 @@ class OperationCommandTest {
         val second = 2
         val input = createInputOperationCommand(first, second)
 
-        operationCommand = DefaultOperationCommand()
+        operationCommand = DefaultOperationCommand
         val result = operationCommand.operate(input)
         assertThat(result).isNotNull
     }
@@ -31,7 +31,7 @@ class OperationCommandTest {
     @CsvSource(value = ["1,2", "3,4"])
     internal fun `두 개의 인자를 더할 수 있다`(first: Int, second: Int) {
         val input = createInputOperationCommand(first, second)
-        operationCommand = AdditionOperationCommand()
+        operationCommand = AdditionOperationCommand
         val result = operationCommand.operate(input)
         assertThat(result).isEqualTo(first + second)
     }
@@ -40,7 +40,7 @@ class OperationCommandTest {
     @CsvSource(value = ["1,2", "3,4"])
     internal fun `첫 번째 인자에서 두 번째 인자를 뺄 수 있다`(first: Int, second: Int) {
         val input = createInputOperationCommand(first, second)
-        operationCommand = SubtractionOperationCommand()
+        operationCommand = SubtractionOperationCommand
         val result = operationCommand.operate(input)
         assertThat(result).isEqualTo(first - second)
     }
@@ -49,7 +49,7 @@ class OperationCommandTest {
     @CsvSource(value = ["1,2", "3,4", "100,0"])
     internal fun `두 개의 인자를 곱할 수 있다`(first: Int, second: Int) {
         val input = createInputOperationCommand(first, second)
-        operationCommand = MultiplicationOperationCommand()
+        operationCommand = MultiplicationOperationCommand
         val result = operationCommand.operate(input)
         assertThat(result).isEqualTo(first * second)
     }
@@ -58,10 +58,10 @@ class OperationCommandTest {
     @CsvSource(value = ["1,2", "3,4"])
     internal fun `첫 번째 인자에서 두 번째 인자를 나눌 수 있다`(first: Int, second: Int) {
         val input = createInputOperationCommand(first, second)
-        operationCommand = DivisionOperationCommand()
+        operationCommand = DivisionOperationCommand
         val result = operationCommand.operate(input)
         assertThat(result).isEqualTo(first / second)
     }
 
-    fun createInputOperationCommand(first: Int, second: Int) = InputOperationCommand(first, second)
+    private fun createInputOperationCommand(first: Int, second: Int) = InputOperationCommand(first, second)
 }
